pyproject.toml
[project]
name = "multidialectalknowledgeextraction"
version = "0.1.0"
description = "Multidialect ASR → Topic → Knowledge Graph (TR/KMR/ZZA)"
authors = [{name="Berfin Çelik"}]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "torch",
  "torchaudio",
  "transformers",
  "datasets",
  "accelerate",
  "evaluate",
  "peft",
  "soundfile",
  "librosa",
  "spacy",
  "sentence-transformers",
  "jiwer",
  "tqdm",
  "matplotlib",
  "pandas",
  "pyyaml",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
addopts = "-q"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
